Remove allowed_restricted_tools (Step 3 not implemented), add skills docs
- Remove allowed_restricted_tools from Skill model, skill_loader,
  skills_fetcher — was added prematurely for Step 3 (per-skill
  restricted tool gating) which hasn't been implemented yet
- Remove metadata field from StructuredToolResult (same reason)
- Add docs/reference/skills.md with migration guide from runbooks,
  SKILL.md format, and configuration instructions
- Migrate CLAUDE.md and README.md from runbooks/ to skills/ with
  minimal changes (word rename + updated paths/format)
- Remove old holmes/plugins/runbooks/ module entirely
- Update docs nav: Runbooks -> Skills
