+++ docs/reference/.nav.yml
@@ -6,6 +6,6 @@ nav:
   - Microsoft Teams Bot Permissions: teams-permissions.md
   - HTTP API: http-api.md
   - OpenTelemetry: opentelemetry.md
-  - Runbooks: runbooks.md
+  - Skills: skills.md
   - Slash Commands: slash-commands.md
   - Troubleshooting: troubleshooting.md
