--- holmes/plugins/runbooks/catalog.json
@@ -1,3 +1,0 @@
-{
-  "catalog": []
-}
